fix(tests): mirror src modules so PS-204 stops flagging orphan tests
CI failed on three PS-204 §2 orphan-test-file errors, all three from this
branch: the tests were named after the BEHAVIOUR they check rather than the
module they mirror, so the auditor could not pair them with a src file.

  test_account_refresh_quota_cache_no_accounts.py -> test__account_refresh_quota_cache.py
  test_pane_context_log_level.py                  -> test__pane_context_log.py
  test_pick_healthy_blind_remedy.py               -> merged into test__pick_healthy.py

The third had nowhere to be renamed to: test__pick_healthy.py already exists,
and PS-204 wants one test file per src module, so its seven blind-refusal
tests move into that file rather than sitting beside it.

The destination carries an autouse fixture that redirects the DEFAULT quota
cache path. These tests pass quota_cache_path= explicitly and so are
unaffected, but that is asserted by running them, not by reading them: all 54
pass, and audit-project now reports no project-structure violations.
